services/app/requirements.txt
amqp==2.4.2
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.2
Django==2.1.7
django-rest-framework==0.1.0
djangorestframework==3.9.2
entrypoints==0.3
factory-boy==2.11.1
Faker==1.0.2
flake8==3.7.7
idna==2.8
inflection==0.3.1
ipdb==0.11
ipython==7.3.0
ipython-genutils==0.2.0
jedi==0.13.3
kombu==4.4.0
mccabe==0.6.1
more-itertools==6.0.0
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pluggy==0.9.0
prompt-toolkit==2.0.9
psycopg2==2.7.7
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.3.1
pytest==4.3.0
pytest-django==3.4.8
pytest-factoryboy==2.0.2
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
requests-mock==1.5.2
six==1.12.0
text-unidecode==1.2
traitlets==4.3.2
urllib3==1.24.1
vine==1.2.0
wcwidth==0.1.7
